+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### v1.0.0-preview.13 - 05 November 2025
+* Fix: When running from named configuration, the name wasn't forwarded to plugin analysis
+
 ### v1.0.0-preview.12 - 05 November 2025
 * Add: Configuration validation and list commands
 * Add: CLI overrides for the configuration based root command